preprocessing/src/file.py
from google.cloud import storage,logging as cloud_logging
import os
from pathlib import Path
from src.logger_config import get_logger

"""
File handling module for the GRAID preprocessing pipeline.
This module provides utilities for cloud storage operations,
local file management, and directory structure creation for
video processing workflows.

Implementation details are masked for security and intellectual property protection.
"""

logger = get_logger()

# Configure storage client
BUCKET_NAME = os.getenv('BUCKET_NAME')
storage_client = storage.Client()
bucket = storage_client.bucket(BUCKET_NAME)

def get_local_video_path(bucket_video_path: str) -> str:
    """
    Downloads a video from Google Cloud Storage to a local temporary directory.
    
    This function:
    1. Creates a local temporary directory if it doesn't exist
    2. Constructs the local file path for the downloaded video
    3. Downloads the video blob from the configured bucket
    4. Logs the download location
    
    Args:
        bucket_video_path (str): Path to the video in the Google Cloud Storage bucket
        
    Returns:
        str: Local file path where the video has been downloaded
        
    Implementation details masked for security and intellectual property protection.
    """
    # Implementation masked for security
    pass

def create_files_in_bucket_from_df(df, path):
    """
    Uploads a pandas DataFrame as a CSV file to Google Cloud Storage.
    
    This function:
    1. Converts the DataFrame to CSV format
    2. Creates a blob at the specified path
    3. Uploads the CSV data to the blob
    4. Logs the successful upload
    
    Args:
        df: Pandas DataFrame to upload
        path: Destination path in the Google Cloud Storage bucket
        
    Implementation details masked for security and intellectual property protection.
    """
    # Implementation masked for security
    pass

def create_directory_structure(base_path: str, video_id: str) -> dict:
    """
    Creates an organized directory structure for video processing.
    
    This function:
    1. Constructs paths for various processing stages (chunks, loader data, etc.)
    2. Creates all required directories if they don't exist
    3. Defines paths for output files
    4. Determines if processing can be skipped based on existing outputs
    
    Args:
        base_path (str): Base directory path
        video_id (str): Unique identifier for the video
        
    Returns:
        dict: Dictionary containing all relevant paths for the video processing pipeline
        
    Implementation details masked for security and intellectual property protection.
    """
    # Implementation masked for security
    pass

def extract_course_video_info_from_path(path):
    """
    Extracts course_id and video_id from a file path.
    
    This function:
    1. Parses a file path like '/storage/1/video/1.mp4'
    2. Extracts the course ID from the path components
    3. Extracts the video ID from the filename without extension
    4. Handles exceptions with appropriate logging
    
    Args:
        path (str): Path containing course and video information
        
    Returns:
        tuple: (course_id, video_id) extracted from the path
        
    Implementation details masked for security and intellectual property protection.
    """
    # Implementation masked for security
    pass