preprocessing/pyproject.toml
[project]
name = "preprocessing"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikhileshwarreddybommareddy",email = "bnikhileshwarreddy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "functions-framework (==3.*)",
    "google-cloud-storage (==2.*)",
    "google-cloud-logging (==2.*)",
    "groq (==0.18.0)",
    "llama-index (==0.12.16)",
    "llama-index-llms-groq (==0.3.1)",
    "llama-index-vector-stores-pinecone (==0.4.4)",
    "llama-index-embeddings-huggingface (>=0.5.3,<0.6.0)",
    "pandas (==2.2.3)",
    "pathlib (>=1.0.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydub (==0.25.1)",
    "openai (>=1.72.0,<2.0.0)",
    "asyncio (>=3.4.3)",
    "requests (>=2.31.0)",
    "pinecone-client (==5.0.1)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
